requirements.txt
# Core Framework
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6

# Blockchain
web3==7.8.0
slither-analyzer==0.9.6
eth-account==0.13.5
eth-utils==5.2.0

# Task Queue
celery==5.3.6
redis==5.2.1
redis-server==6.0.9

# ML Dependencies
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.0.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.11.13
websockets==13.1
logging==0.4.9.6

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
